Support in Libra for Synthetic Minds' Move Verifier and Synthesizer
Details of this work done have been described in a blog posts on:
* Design: https://synthetic-minds.com/pages/blog/blog-2019-09-10.html
* Verification: https://synthetic-minds.com/pages/blog/blog-2019-09-11.html
* Synthesis: https://synthetic-minds.com/pages/blog/blog-2019-09-12.html

At Synthetic Minds we build technology to accelerate the development of
correct smart contracts. We retargeted our verifier to Move and then
wrote Move contracts and libraries, observing how Move’s linear type
system and our verifier stacked up in ensuring correctness of Move code.

These code changes represent the following changes: port the verifier
and building illustrative Move contracts, which include a Time
Service, Name Server, MultiSig wallet, and a toy example that forms the
basis of Escrow and Auction.

We also verified properties of the stdlib LibraCoin and LibraAccount
modules that are described in the blog post.

Synthetic Minds' Move Verifier and Synthesizer is now operational and
allows us to validate the correctness of Move modules simply by taking
specs from test scenarios, which should make verification as accessible
as testing. Additional specifications can be added in the form of
assertions.

In order to allow Synthetic Minds' verifier to operate on .mvir files we
added the ability to output a json representation of the AST.

We are including test scenarios (as transaction scripts). These test
scenarios are the correctness conditions for the verifier. We also
include generated scripts that check correctness on the concrete
testnet.

Note: Due to limited support for user-defined modules when we forked the
repository, we created our modules and demo application in stdlib/